+++ ant-node/src/bin/antnode/main.rs
@@ -27,7 +27,7 @@ use ant_evm::{EvmNetwork, RewardsAddress, get_evm_network};
 use ant_logging::metrics::init_metrics;
 use ant_logging::{Level, LogFormat, LogOutputDest, ReloadHandle};
 use ant_node::utils::{get_antnode_root_dir, get_root_dir_and_keypair};
-use ant_node::{Marker, NodeBuilder, NodeEvent, NodeEventsReceiver, RunningNode};
+use ant_node::{Marker, NodeBuilder, NodeEvent, NodeEventsReceiver};
 use ant_protocol::{
     node_rpc::{NodeCtrl, StopResult},
     version,
@@ -38,9 +38,7 @@ use const_hex::traits::FromHex;
 use libp2p::PeerId;
 use rand::Rng;
 use std::{
-    collections::hash_map::DefaultHasher,
     env,
-    hash::{Hash, Hasher},
     net::{IpAddr, Ipv4Addr, SocketAddr},
     path::PathBuf,
     process::Command,
@@ -52,10 +50,6 @@ use tokio::{
     time::sleep,
 };
 use tracing_appender::non_blocking::WorkerGuard;
-
-/// Default network size estimate used when the actual size cannot be determined.
-/// This represents a baseline network of 100,000 nodes.
-const DEFAULT_NETWORK_SIZE: usize = 100_000;
 
 #[derive(Debug, Clone)]
 pub enum LogOutputDestArg {
@@ -238,46 +232,6 @@ struct Opt {
     /// Useful when running under a service manager that handles restarts.
     #[clap(long, default_value_t = false)]
     stop_on_upgrade: bool,
-}
-
-/// Calculates a deterministic restart delay based on peer ID and network size.
-///
-/// This algorithm ensures that node upgrades are staggered across the network to prevent
-/// many nodes from restarting simultaneously.
-///
-/// The delay is calculated as follows:
-/// 1. Hash the node's peer ID to get a deterministic value unique to this node
-/// 2. Calculate a time range in hours based on network size: min(72, network_size/100_000 + 1)
-///    - Smaller networks (< 100k nodes) get 1-2 hour windows
-///    - Larger networks get progressively longer windows, capped at 72 hours (3 days)
-/// 3. Use the hash modulo the time range to assign this node a specific delay
-///
-/// This approach provides:
-/// - Deterministic delays: Same peer ID always gets the same delay for a given network size
-/// - Even distribution: Hash modulo spreads nodes uniformly across the time window
-/// - Network-aware scaling: Larger networks get longer upgrade windows
-///
-/// # Arguments
-/// * `running_node` - The node to calculate the restart delay for
-///
-/// # Returns
-/// A `Duration` representing when this node should restart for an upgrade
-async fn calculate_restart_delay(running_node: &RunningNode) -> Duration {
-    let peer_id = running_node.peer_id();
-    let est_network_size = running_node
-        .get_estimated_network_size()
-        .await
-        .unwrap_or(DEFAULT_NETWORK_SIZE);
-
-    let mut hasher = DefaultHasher::new();
-    peer_id.hash(&mut hasher);
-    let hash_value = hasher.finish();
-
-    let time_range_hours = std::cmp::min(72, (est_network_size / DEFAULT_NETWORK_SIZE) + 1);
-    let time_range_seconds = time_range_hours * 3600;
-    let upgrade_time_seconds = (hash_value as usize) % time_range_seconds;
-
-    Duration::from_secs(upgrade_time_seconds as u64)
 }
 
 fn main() -> Result<()> {
@@ -499,13 +453,13 @@ You can check your reward balance by running:
         let ctrl_tx_restart = ctrl_tx.clone();
         let running_node_clone = running_node.clone();
         tokio::spawn(async move {
+            // 72 hours (259200 seconds) with ±5% randomization to prevent simultaneous upgrades
+            let base_delay = 259200;
+            let variance = rand::thread_rng().gen_range(-12960..=12960);
+            let upgrade_check_delay_secs = (base_delay + variance) as u64;
+            let upgrade_check_wake_time =
+                chrono::Utc::now() + chrono::Duration::seconds(upgrade_check_delay_secs as i64);
             loop {
-                // 72 hours (259200 seconds) with ±5% randomization to prevent simultaneous upgrades
-                let base_delay = 259200;
-                let variance = rand::thread_rng().gen_range(-12960..=12960);
-                let upgrade_check_delay_secs = (base_delay + variance) as u64;
-                let upgrade_check_wake_time =
-                    chrono::Utc::now() + chrono::Duration::seconds(upgrade_check_delay_secs as i64);
                 info!(
                     "Next upgrade check scheduled for {}",
                     upgrade_check_wake_time
@@ -514,7 +468,8 @@ You can check your reward balance by running:
 
                 match upgrade::perform_upgrade().await {
                     Ok(()) => {
-                        let node_restart_delay = calculate_restart_delay(&running_node_clone).await;
+                        let node_restart_delay =
+                            upgrade::calculate_restart_delay(&running_node_clone).await;
                         let node_restart_wake_time = chrono::Utc::now() + node_restart_delay;
                         info!("Node will stop/restart for upgrade at {node_restart_wake_time}");
                         sleep(node_restart_delay).await;
